chat_bot file update
